ai_service/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Google Gemini AI
google-generativeai==0.3.2

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Data Processing
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-jose[cryptography]==3.3.0

# Logging
structlog==24.1.0

